requirements.txt
# Core dependencies for Algorithmic Market Making MVP
numpy>=1.21.0
pandas>=1.3.0
pyyaml>=6.0
scikit-learn>=1.0.0
scipy>=1.7.0

# Optional dependencies for advanced features
# statsmodels>=0.13.0  # Uncomment for advanced cointegration analysis
# matplotlib>=3.5.0    # Uncomment for plotting capabilities
# jupyter>=1.0.0       # Uncomment for notebook analysis

# Development dependencies (optional)
# pytest>=6.0.0        # For running tests
# black>=22.0.0        # For code formatting
# flake8>=4.0.0        # For linting
